+++ lib/tree/binarymaxheap.rb
@@ -38,6 +38,7 @@
 # frozen_string_literal: true
 
 require_relative 'binarytree'
+require_relative 'utils/heap_shared_methods'
 
 module Tree
   # Provides a Binary Max-Heap implementation. This node allows only two child
@@ -56,6 +57,8 @@ module Tree
   # This inherits from the {Tree::BinaryTreeNode} class.
   #
   class BinaryMaxHeapNode < BinaryTreeNode
+    include Tree::Utils::HeapSharedMethods
+
     # Inserts the specified node into the heap, based on the node content.
     #
     # @param [Tree::BinaryMaxHeapNode, String, Symbol] node_or_name The node
@@ -224,24 +227,6 @@ module Tree
       raise ArgumentError, 'Binary max-heap key (content) must not be nil.' if key.nil?
     end
 
-    # +true+ if the root represents an empty heap.
-    #
-    # @param [Tree::BinaryMaxHeapNode] node The node to check.
-    # @return [Boolean] +true+ if empty.
-    def empty_root?(node)
-      node.root? && node.content.nil? && !node.children?
-    end
-
-    # Returns the heap size treating an empty root as size 0.
-    #
-    # @param [Tree::BinaryMaxHeapNode] node The root node.
-    # @return [Integer] The heap size.
-    def heap_size(node)
-      return 0 if empty_root?(node)
-
-      node.size
-    end
-
     # Insert a node at the next available index to keep the heap complete.
     #
     # @param [Tree::BinaryMaxHeapNode] tree_root The root node.
@@ -258,31 +243,6 @@ module Tree
       else
         parent.right_child = node
       end
-    end
-
-    # Return the node at the given heap index.
-    #
-    # @param [Tree::BinaryMaxHeapNode] tree_root The root node.
-    # @param [Integer] index The heap index (1-based).
-    # @return [Tree::BinaryMaxHeapNode, nil] The node at the index.
-    def node_for_index(tree_root, index)
-      return tree_root if index == 1
-
-      current = tree_root
-      path_bits = index.to_s(2).chars.drop(1)
-      path_bits.each do |bit|
-        current = bit == '0' ? current.left_child : current.right_child
-        return nil unless current
-      end
-      current
-    end
-
-    # Return the last node in the heap (right-most node at the deepest level).
-    #
-    # @param [Tree::BinaryMaxHeapNode] tree_root The root node.
-    # @return [Tree::BinaryMaxHeapNode] The last node.
-    def last_node(tree_root)
-      node_for_index(tree_root, heap_size(tree_root))
     end
 
     # Restore heap ordering by bubbling the node upward.
@@ -317,56 +277,6 @@ module Tree
         swap_contents(current, swap_candidate)
         current = swap_candidate
       end
-    end
-
-    # Swap the content values of two nodes.
-    #
-    # @param [Tree::BinaryMaxHeapNode] left The first node.
-    # @param [Tree::BinaryMaxHeapNode] right The second node.
-    # @return [void]
-    def swap_contents(left, right)
-      left.content, right.content = right.content, left.content
-    end
-
-    # Remove the last node from the heap.
-    #
-    # @param [Tree::BinaryMaxHeapNode] node The last node.
-    # @return [void]
-    def remove_last_node(node)
-      if node.root?
-        node.instance_variable_set(:@content, nil)
-        return
-      end
-
-      if node.left_child?
-        node.parent.left_child = nil
-      else
-        node.parent.right_child = nil
-      end
-    end
-
-    # Remove an arbitrary node from the heap.
-    #
-    # @param [Tree::BinaryMaxHeapNode] node The node to remove.
-    # @return [void]
-    def remove_node(node)
-      tree_root = root
-      return if empty_root?(tree_root)
-
-      if heap_size(tree_root) == 1
-        tree_root.instance_variable_set(:@content, nil)
-        return
-      end
-
-      last = last_node(tree_root)
-      if node.equal?(last)
-        remove_last_node(last)
-        return
-      end
-
-      swap_contents(node, last)
-      remove_last_node(last)
-      rebalance_after_swap(node)
     end
 
     # Rebalance heap ordering after swapping in a replacement node.